requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
PyQt6==6.7.1
pytest==8.3.3
httpx==0.27.2
pyinstaller==6.10.0
